Rename `gh extensions` → `gh extension`
This is for compatibility with other core commands which are all singular.
