+++ pkg/cmd/extension/extension.go
@@ -1,4 +1,4 @@
-package extensions
+package extension
 
 import (
 	"path/filepath"
