+++ pkg/cmd/extension/manager.go
@@ -1,4 +1,4 @@
-package extensions
+package extension
 
 import (
 	"bytes"
