+++ pkg/cmd/extension/manager_test.go
@@ -1,4 +1,4 @@
-package extensions
+package extension
 
 import (
 	"bytes"
