+++ pkg/cmd/extension/symlink_windows.go
@@ -1,4 +1,4 @@
-package extensions
+package extension
 
 import "os"
